minor repackaging
